feat: add mock CallSid simulation support
