+++ lib/mock/twilio/decorator.rb
@@ -19,7 +19,7 @@ module Mock
 
           schema = url_from(body, request)
           # return body decorate if needed
-          return ENDPOINTS[schema].decorate(body, request) if schema
+          return ENDPOINTS[schema].for(body, request) if schema
 
           body
         end
