+++ lib/mock/twilio/schemas/api_2010.rb
@@ -1,10 +1,18 @@
 # frozen_string_literal: true
 
+require_relative "../decorators/api_2010/messages"
+require_relative "../decorators/api_2010/calls"
+
 module Mock
   module Twilio
     module Schemas
       class Api2010
         class << self
+          RESOURCES = {
+            messages: Mock::Twilio::Decorators::Api2010::Messages,
+            calls: Mock::Twilio::Decorators::Api2010::Calls,
+          }
+
           PAGES_KEYS = [
             "end",
             "first_page_uri",
@@ -19,35 +27,15 @@ module Mock
             "uri"
           ].freeze
 
-          def decorate(body, request)
-            body["date_updated"] = Time.current.rfc2822 if body["date_updated"]
-            body["date_sent"] = Time.current.rfc2822 if body["date_sent"]
-            body["date_created"] = Time.current.rfc2822 if body["date_created"]
-            body["start_time"] = Time.current.rfc2822 if body["start_time"]
-            body["end_time"] = Time.current.rfc2822 if body["end_time"]
+          def for(body, request)
+            url = request.url.split(request.host).last
 
-            message_sid(body, request) if body["sid"]
-            pagination(body) if body["available_phone_numbers"]
-
-            body
-          end
-
-          def pagination(body)
-            # Params returned in mock_server but not on real twilio request for the moment.
-            # Not needed for us now.
-            PAGES_KEYS.each do |key|
-              body.delete(key) if body[key]
+            case url
+            when %r{\/2010-04-01/Accounts/[A-Za-z0-9]+/Messages.json}
+              RESOURCES[:messages].decorate(body, request)
+            when %r{\/2010-04-01/Accounts/[A-Za-z0-9]+/Calls.json}
+              RESOURCES[:calls].decorate(body, request)
             end
-          end
-
-          def message_sid(body, request)
-            prefix = request.data["MediaUrl"] ? "MM" : "SM"
-            sid = prefix + SecureRandom.hex(16)
-            scheduler = Rufus::Scheduler.new
-            scheduler.in '2s' do
-              Mock::Twilio::Webhooks::Messages.trigger(sid)
-            end
-            body["sid"] = sid
           end
         end
       end
